Added project detail in API
One of the features that I'm hoping to eventually use is the ability
to connect Tock data with other data. Given that Tock stores some
descriptive metadata, I wanted to be able to access that data at the project
level. In other words, this PR would expose the following endpoint:

`/api/projects/:number.json`