Closes #267:  Add support for entry notes (snippets)
This changeset introduces the ability to add notes to a time entry. These notes can be optional or required and applied to any project. In order to enable them, go into the Django admin and find a Project that you would like to enable notes on. In the admin form you will now see to additional checkboxes to enable the display of a notes field, and wether or not it is required with time entry.

A new dependency was added to the project to support this enhancement, called bleach (https://bleach.readthedocs.org/). This was added per the Django documentation note on using the striptags template filter: https://docs.djangoproject.com/en/1.9/ref/templates/builtins/\#striptags

A couple of tests have been added and a couple of others have been updated to support this new enhancement.  In addition, the API output has been updated to also include any notes entered by users and several Django admin features have been enabled to make it easier to work in.