wrote a FormView and its associated HTML template, and added its URL route
The form template is an extension of the `base.html` template, and may be accessed at `/employees/<username>/travel_request` provided that the currently logged-in user is either `<username>` or a superuser.  (I also added this link to the `user_form.html` template for the sake of convenience.)

The form template employs [CSRF protection](https://docs.djangoproject.com/en/1.9/ref/csrf/#how-to-use-it).

The `requestor_name` and `requestor_email` fields are automatically populated with the full name and email address, respectively, of the user corresponding to `<username>`.

The form template uses the jQuery Datepicker widget, for the sake of [consistency](https://github.com/18F/tock/blob/master/tock/tock/templates/employees/user_form.html#L29).