added a reminder for users to submit their timesheets
When a user saves (but not submits) a timesheet, she is presented with a flash message which both confirms that her timesheet has been saved and reminds her to submit her timesheet when she has finished her edits.

To facilitate styling, I created an `alert` mixin with an optional `$color` argument.  The flash message is presented in yellow, to contrast with the static `alert-red` message.  I then recompiled the stylesheets.