wrote unit test
The test ensures that the total number of hours for the project whose `id` is `1` matches what's displayed on the project details page.  (Note, however, that this particular project is the only one in the test fixtures for which any hours have been entered.)

To facilitate testing, I added an `id` attribute to the `<span>` element containing the total number of hours.