Update README to reflect native Docker apps for Mac and Windows
This resolves #440 by updating the README to reflect that Docker
now has native versions for both Mac and Windows.

docker-compose is automatically installed and docker-machine is
no longer necessary because of Docker's integration with the OS.

This peripherally resolves #503 by adding an extra step in the
Getting Started section to clarify that the IP address from the
app container can be ignored (since it's behind nginx anyway).