Remove Travis CI integration
This changeset removes the Travis CI configuration as the first step in switching to Circle CI, which is approved for 18F GitHub repositories.  Travis CI is no longer approved and we must turn off integration with it.

Once this is merged in, an admin will need to go in and remove the integration within Travis CI and/or GitHub itself.