Fix ZeroDivisionError when accessing some timecards
This changeset fixes the ZeroDivisionError that sometimes gets thrown when accessing certain timecards.  It appears that this happens when the `work_hours` end up being 0 when retrieving all Float tasks for a user.  It is unclear how to fix this at the moment, but one thing that jumped out was the tests around this functionality were all lumped into a single test.  To start addressing this, I pulled apart all of the different tests to exercise the functionality separately so that we can then add another test for this specific scenario.  We will need to determine what the appropriate course of action or information to display is if the `work_hours` calculation ends up yielding 0.