Handling Float API errors gracefully
This changeset provides much more robust error handling for the Float API calls and more graceful handling of them so that users can still see and enter their timecards.  If an error occurs while accessing Float data, the user is presented with a message saying so and the error is logged behind the scenes so we can see what went wrong and why.  We currently only see a stack trace saying that we cannot call a method on a NoneType object.