Remove hardcoded note and fix reporting period-specific messages
This changeset finalizes the new timecard notes feature by removing the hardcoded message at the top of the timecard form and making it an initial model available for immediate use.  It also fixes the display of the reporting-period specific messages and provides a few extra options for them as well:

* An optional title
* A style to choose from (same as timecard notes)
* Whether or not the message is actually enabled for display

These messages will appear above any general messages as they are reporting period specific if they are enabled.  If they are not, they can still be useful for an admin for internal notes.