Update Docker and Python project dependencies
This changeset gets Tock more up to date with more recent Docker container versions and Python package updates.  This will be required for our ATO work and some of the Python packages will have to be updated piecemeal to account for testing in the interim to make sure things are still working okay.