Add helper methods for filtering projects
This changeset adds a `ProjectManager` model manager (not a human being...) for the `Project` model that provides a couple of convenience methods for easily retrieving all active and inactive projects.  The Project admin page already has the filtering exposed for inactive and active projects, but this will be needed when we start adding support to pre-populate time cards and only want to work through a list of active projects.