Add support for timecard prefill data
This changeset adds support for managing staff assignments to projects in order to prefill timecards for users.  It adds a new model that tracks the assignment of a user to a project, the number of hours they are assigned, and any additional notes about the particular assignment.  Timecards will take this data and prefill existing entries with the information, adding the Tock lines (projects) if need be but also preserving the existing functionality of loading all Tock lines from the previous week.