Add handbook page link and update Python to 3.6.4
This changeset adds a link to the Tock handbook page in the README and the timecard form itself.  In the case of the timecard form, it is a hardcoded info message that will appear below every other message to ensure that important and especially critical things are not missed.

CircleCI and cloud.gov have recently updated their container/buildpack versions as well, which means we can use Python 3.6.4.  This changeset accounts for that as well.