Add support for additional prefill fields
This changeset adds support for three new fields on the prefill data:  is_active, start_date, and end_date.  The is_active field is fully integrated and accounted for; only active prefill data records will be used when viewing a timecard.  This ensures two things:

* These records can stick around per federal record-keeping guidelines
* Records can be verified before being enabled

The start_date and end_date fields are optional and could be used to set an automatic window for when a prefill record should be used by Tock.  This will be fully supported by future work (issue to be written).