Add bandit and run it in CI.
Bandit is a security-focused linter. It found lots of subprocess-related
concerns in manage.py and docker_entrypoint.py, but that's not (directly) used
in production so it seems like it's worth ignore. If this shortcut isn't used
frequently, it might make sense to remove it (as manage.py _is_ referenced in
production).