Add support for additional non-billable projects for the snippets report
This changeset adds a bit of flexibility to the snippets report to include additional project line items instead of just the old General line item.  The reason for this is once the General project was retired in favor of the newer non-billable project line items, the report lost its usefulness and would not display any of the new information.  If we ever want to add additional projects to this report, it should just be a matter of including the project ID with the newly defined constant.