Add support for Markdown formatting and timecard notes ordering
This changeset adds support for Markdown formatting of Reporting Period messages and Timecard Notes.  It also adds support for ordering the Timecard Notes if you have more than one; it defaults to setting the position to be 1 more than the current number of existing Timecard Notes (regardless of whether or not they are enabled).