Update and Move vendored dependencies
With this commit we move all of our vendored js/css
dependencies into a single folder. We're also renaming
w/ versions and making the necessary template changes
to update paths.

This will give us greater visibilty into the current versions
being served by tock.

As part of this cleanup, remove datatables.js entirely
as it is not currently in use.