State only migration to sync project state
In 762cbe104c86ffed72b9bd4b6932470a8c194ed7
we manually update the database but did not
also update the Django project state.

This lead to a divergence in database and project state.

Specifically, while a unique constraint exists in the database
after hours.0046. Django is not aware of it, recognizes the
the difference between the code and the project state,
and attempts to recreate the unique constraint with a
new migration.

This new migration makes no database changes while updating
the project state to accurately reflect the existance of the
`hours_timecardobject` unique constraint.