Removed jQuery from getFormData()
I think I need more tests for this. That is, I need tests for all of the
fields of the entries, as well as a test that exercises whether the
deletion filter is working. (The function getFormData does a few things,
and therefore I might split it up for further testing. To be seen.)