Sequential tests, fixed.
First, I've added "runInBand" to the jest config. This runs tests
sequentially. I have deep suspicions that cookies are stored in the
puppeteer Chromium instance, and that is causing problems for async
tests. Therefore, everything runs one at a time.

This was also a problem in the test suite... after running a suite of
tests, it is necessary to hit "/logout" before running the next suite.
Using the "afterAll()" function of puppeteer (or, Jest... I think it was
Jest, actually) to teardown after each suite was necessary.

Suites renamed. This is a side-effect of me trying to force an ordering.
This did not work.