Concurrency and state must be noted in jest tests
This is a set of tests that passes for the page interactions. The
challenge is that tock maintains some state in the browser context, and
as a result, there are interactions between sessions that must be
remembered.

For example, adding a timecard entry is persisted in the local state. As
a result, the headless Chromium instance will hold on to state as well.
Elements added during one test will exist in the next.

I've navigated around this; now, I'm going to look at clearing the
state, so that each test can be clean. That will make this a lot easier.