Successfully deleted items in jest/puppeteer.
Checkboxes are bad. We should not use them in 18F applications.

Because DOM interactions with checkboxes are not ... the same as other
Javascript/DOM interactions, we need to click the label element, not the
checkbox itself. Attempts to set the "checked" state of the checkbox
yielded little success, even with some radical gymnastics to access the
DOM state. Someone with deep Puppeteer/Jest/JS/DOM knowledge may know
Exactly What To Do, but I do not.

The solution ultimately landed on in order to 1) add two projects with
time, 2) click the delete checkbox on the first, and 3) verify that the
getFormData() function ignores that first item was to *click the label*
for the checkbox.

So. Testing continues. This was a blocker.