Removed a page traversal in populateHourTotals()
This function did a page traversal. In doing so, it walked all the
entries, pulling the project (from the select drop down) and the hours
(from the text field). This is the same thing that getFormData() does.

I removed the traversal (which used jQuery) and replaced it with a call
to getFormData(). Because this was stored to local storage, I then
filtered the result from getFormData() so that only the project and
hours values were kept in an array of objects. This made it equivalent
to the previous traversal.

The previous commit added a test to make sure that localStorage had the
correct state after a call to populateHourTotals(). The test passed
before and after.

Note the singular on "test". I may add more tests. However, the change
is now made, which means any additional tests will simply test the
functionality as-is. Hm.