Fixed addition, removal of entries.
This is now vanilla JS.

I don't know that I like the way we clone an existing entry box and then
empty it. I feel like this is error prone.

I would like it if the boxes were an include in Jinja, and the include
was used to render the boxes in the backend. That way, we could also
render one, blank, invisible template, and use the invisible one to
create new, empty elements.

We still have numbering challenges, but that's... a problem for another
day as well. I'm unclear why we need to number these... that is, could
we just generate random IDs (or not use IDs at all), and still get the
desired behavior? I don't know... but, I think the numbering might be a
red herring, and we could simplify this further. Something to
discuss/consider.