Allow running Jest watch mode in Docker
When the `CI` environment variable is set, Jest assumes you're not in an interactive session, which makes watch mode kind of difficult to use. This commit copies some of the de-sandboxing jest-puppeteer does in CI environments and applies it when it detects it is running in Docker, which lets watch mode run interactively.