Add element IDs and show/hide function
This adds two new IDs to the hourly sum elements in the footer so
that they can be cleanly selected without relying on the class of
`grid-col-2` which is a layout/UI structure class that could
change in the future to support different viewport sizes/devices.

It also adds a method that wraps up the functionality of needing
to show or hide the hourly sum elements depending on the presence
of a weekly billed project existing on the timecard.