Fix server error and hour validation for weekly billing
When attempting to submit a timecard from the admin interface if
you left the "Hours Spent" column completely blank and didn't
enter a zero (0) in the process of adding a weekly billing
project it would cause the server to throw a TypeError. There is
now a guard statement in the admin code to prevent this.

Additionally, the hourly validation logic was still being applied
to TimecardObjects even if there was a weekly billing project on
the Timecard. Now we don't throw the hourly validation warning
if a weekly billing project is detected on the Timecard.